+++ packages/snap-client/src/Client/transforms/index.ts
@@ -1,3 +1,5 @@
 export * from './searchRequest';
 export * from './searchResponse';
 export * from './suggestResponse';
+export * from './recommendationFiltersGet';
+export * from './recommendationFiltersPost';