+++ packages/snap-preact/src/Instantiators/RecommendationInstantiator.tsx
@@ -159,7 +159,8 @@ export class RecommendationInstantiator {
 							branch?: string;
 							batched?: boolean;
 							realtime?: boolean;
-							categories?: any;
+							categories?: string[];
+							brands?: string[];
 							limit?: number;
 						}
 					}
@@ -179,6 +180,12 @@ export class RecommendationInstantiator {
 				}
 				if (options?.categories) {
 					contextGlobals.categories = options.categories;
+				}
+				if (options?.filters) {
+					contextGlobals.filters = options.filters;
+				}
+				if (options?.brands) {
+					contextGlobals.brands = options.brands;
 				}
 				if (options?.limit && Number.isInteger(Number(options?.limit))) {
 					contextGlobals.limits = Number(options?.limit);
@@ -206,7 +213,10 @@ export class RecommendationInstantiator {
 				const defaultGlobals = {
 					limits: 20,
 				};
-				const globals = deepmerge(deepmerge(defaultGlobals, this.config.client?.globals || {}), contextGlobals);
+				const globals = deepmerge(
+					deepmerge(deepmerge(defaultGlobals, this.config.client?.globals || {}), (this.config.config?.globals as any) || {}),
+					contextGlobals
+				);
 
 				const controllerConfig = {
 					id: `recommend_${tag}_${profileCount[tag] - 1}`,