+++ packages/snap-controller/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.49.0](https://github.com/searchspring/snap/compare/v0.48.0...v0.49.0) (2023-10-19)
+
+### Bug Fixes
+
+- add options.brands to recommendation instantiator context ([afb6c4d](https://github.com/searchspring/snap/commit/afb6c4d19730fc3e120cc3c9e941a3d223373da8))
+
+### Features
+
+- **client-recommend:** adding filtering to recommendations ([d711f7b](https://github.com/searchspring/snap/commit/d711f7b79237132ec89dfc33c18c2fe3242a1503))
+
+# [0.48.0](https://github.com/searchspring/snap/compare/v0.47.0...v0.48.0) (2023-10-11)
+
+### Bug Fixes
+
+- **controller/finder:** altering the beforeFind middleware to ensure redirection is blocked ([936e184](https://github.com/searchspring/snap/commit/936e184ba932fce75b379eeb6e197af0648aac68))
+
+### Features
+
+- add autocomplete settings.redirects.singleResult config ([e620701](https://github.com/searchspring/snap/commit/e620701cb7d528ea4d3fe823bd0104e1bdd73a31))
+
 # [0.47.0](https://github.com/searchspring/snap/compare/v0.46.0...v0.47.0) (2023-09-15)
 
 ### Features